refactor: extract GuzzleFactory. Extract Response class
