+++ lib/ChargeBee/Response.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace ChargeBee\ChargeBee;
+
+class Response {
+    public $data;
+    public $headers;
+
+    public function __construct($data, $headers) {
+        $this->data = $data;
+        $this->headers = $headers;
+    }
+}
